Allow EXTERNAL as well as AMQPLAIN auth
I'm trying to publish messages from openQA (which uses this
client library) to fedora-messaging (which is a RabbitMQ broker
using EXTERNAL auth), so I kinda need this. The intent here is
that by default we'll try and use either of AMQPLAIN or EXTERNAL
if the server advertises one or both, preferring AMQPLAIN; the
auth_mechanism query param can be used to specify only one or
the other.

Signed-off-by: Adam Williamson <awilliam@redhat.com>